+++ R/compartment_property.R
@@ -61,12 +61,12 @@ setGeneric("getRecordName", function(object) {
 })
 
 #_______________________________________________________________________________
-#----                             replaceAll                                ----
+#----                             replace_all                                ----
 #_______________________________________________________________________________
 
-#' @rdname replaceAll
-setMethod("replaceAll", signature=c("compartment_property", "pattern", "character"), definition=function(object, pattern, replacement, ...) {
-  object@rhs <- object@rhs %>% replaceAll(pattern=pattern, replacement=replacement, ...)
+#' @rdname replace_all
+setMethod("replace_all", signature=c("compartment_property", "pattern", "character"), definition=function(object, pattern, replacement, ...) {
+  object@rhs <- object@rhs %>% replace_all(pattern=pattern, replacement=replacement, ...)
   return(object)
 })