+++ R/model_unknown_statement.R
@@ -38,12 +38,12 @@ setMethod("getName", signature = c("unknown_statement"), definition = function(x
 })
 
 #_______________________________________________________________________________
-#----                             replaceAll                                ----
+#----                             replace_all                                ----
 #_______________________________________________________________________________
 
-#' @rdname replaceAll
-setMethod("replaceAll", signature=c("unknown_statement", "pattern", "character"), definition=function(object, pattern, replacement, ...) {
-  object@line <- object@line %>% replaceAll(pattern=pattern, replacement=replacement, ...)
+#' @rdname replace_all
+setMethod("replace_all", signature=c("unknown_statement", "pattern", "character"), definition=function(object, pattern, replacement, ...) {
+  object@line <- object@line %>% replace_all(pattern=pattern, replacement=replacement, ...)
   return(object)
 })